requirements.txt
accelerate==0.23.0
datasets==2.16.1
deepspeed==0.14.0
huggingface_hub==0.20.2
numpy==1.26.4
peft==0.4.0
PyYAML==6.0.1
PyYAML==6.0.1
tenacity==8.2.3
torch==2.1.2
tqdm==4.66.1
transformers==4.37.0
trl==0.4.7
vllm==0.3.0
rank_bm25==0.2.2